Use unstable cache to cache elastic search result for 30 seconds
